// cg-web/src/types/device.ts
export type DeviceModelAuthorityType = {
  detail?: string;
  edit?: string;
};

export type DeviceModelType = {
  type?: DeviceModelTypeEnum;
  specs?: any;
  name?: string;
};

export type DevicePropsType = {
  id?: string;
  name?: string;
  groupName?: string;
  dataType?: DeviceModelType;
  properties?: DevicePropsType[];
  component?: string;
  showHistory?: boolean;
  sourceId?: string;
};

export type TipType = {
  title?: string;
  content?: string;
};

export type DeviceArrayType = {
  type?: DeviceModelTypeEnum;
  specs?: {
    size?: number;
    item?: DeviceTimeRangeType | DeviceStructType;
    originalItem?: DeviceTimeRangeType | DeviceStructType;
  };
};

export type DeviceTreeSelectType = {
  type?: DeviceModelTypeEnum;
  specs?: {
    id?: string;
    name?: string;
  }[];
};

export type DeviceTimeRangeType = {
  type?: DeviceModelTypeEnum;
  specs?: {
    length: number;
  };
};

export type DeviceEnumType = {
  type?: DeviceModelTypeEnum;
  specs?: Record<string, any>;
};

export type DeviceDoubleType = {
  type?: DeviceModelTypeEnum;
  specs?: {
    unit?: string;
    unitName?: string;
    min?: number;
    max?: number;
    enable: boolean;
  };
};

export type DeviceStructType = {
  type?: DeviceModelTypeEnum;
  specs?: DeviceServiceModelType[];
};

export type DeviceServiceModelType = {
  id?: string;
  deviceId?: string;
  parentId?: string;
  serviceId?: string;
  queryId?: string;
  name?: string;
  required?: boolean;
  span?: number;
  showType?: DeviceModelShowTypeEnum;
  disabled?: string;
  tip?: string;
  authority?: DeviceModelAuthorityType[];
  type?: DeviceModelDescribeTypeEnum;
  dataType?:
  | DeviceArrayType
  | DeviceDoubleType
  | DeviceStructType
  | DeviceTimeRangeType
  | DeviceEnumType
  | DeviceTreeSelectType
  | {
    type?: DeviceModelTypeEnum;
    specs?: DeviceServiceModelType[];
  };
  buttons?: 'refresh' | 'edit';
  defaultValue?: string;
  showHistory?: boolean;
  form?: {
    span?: number;
    showType?: DeviceModelShowTypeEnum;
  };
  promptRule?: TipType[];
  controlIntercept?: string;
};